Add files via upload
Source_target_database.sql

This SQL  file contains a script which creates a database which contains a This script creates a database with sources which contains data and a target which data needs to be filled. Using SQL queries tables in the target filled from the source tables.

--------------------------------------------------------------------------------------------------------------------------------------------------------
HosMang_Database.sql contains a file which creates RDBMS database for a Hospital management.

 The aim of this case study is to design and develop a database for the hospital to maintain the
records of various departments, rooms, and doctors in the hospital. It also maintains records of
the regular patients, patients admitted in the hospital, the check up of patients done by the
doctors, the patients that have been operated, and patients discharged from the hospital.